utils api organization and some cleanup
