+++ src/app/containers/Transaction/Transaction.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import style from './Transaction.css'
 import neoPNG from '../../../img/icon-34.png'
 
-import * as Neoscan from '../../utils/NeoscanApi'
+import * as Neoscan from '../../utils/api/neoscan'
 
 import * as string from '../../utils/string'
